refactor: move Exec and Logging to workload_ecs.go
